mcp-server-backend-tools/requirements.txt
fastapi
fastapi_mcp
uvicorn
pyodbc
azure.storage.blob
selenium
markdownify
Pillow
pillow_heif
requests
base64
